Ну хоть работает

+++ src/Service/ffmpeg.ts
@@ -1,7 +1,9 @@
 import { spawn, spawnSync } from 'child_process';
+import Logger from 'log4js';
 import { PassThrough, Readable } from 'stream';
 
 let ffmpegCmd: string;
+let logger: Logger.Logger | undefined = undefined;
 function findFfmpeg() {
     for (const command of ['ffmpeg', 'avconv', './ffmpeg', './avconv', 'ffmpeg.exe', './ffmpeg.exe']) {
         if (!spawnSync(command, ['-h']).error) {
@@ -15,7 +17,11 @@ function findFfmpeg() {
 }
 findFfmpeg();
 
-function ConvertAudioForDis(stream: Readable): Readable {
+function ConvertAudioForDis(stream: Readable): [Readable, () => void] {
+    if (!logger) {
+        logger = Logger.getLogger('ffmpeg_wrapper');
+    }
+
     const child = spawn(ffmpegCmd, [
         '-i', 'pipe:0',
         '-vn', '-hide_banner',
@@ -28,19 +34,32 @@ function ConvertAudioForDis(stream: Readable): Readable {
     ]);
 
     stream.on('error', (err) => {
-        console.log(`Could not convert stream: ${err}`);
+        logger.debug('Could not convert stream', err);
         stream.unpipe();
-        child.stdout.emit('error', `Could not convert stream: ${err}`);
-        child.kill('SIGKILL');
+        pt.emit('error', err);
+        cleanup();
     });
+    stream.on('end', () => cleanup);
+
+    const cleanup = () => {
+        if (!child.killed) {
+            child.kill('SIGKILL');
+        }
+    };
+
+    child.stdin.on('error', (e) => logger.warn('in', e));
+    child.stdout.on('error', (e) => logger.warn('out', e));
 
     stream.pipe(child.stdin);
-    // return child.stdout;
     const pt = new PassThrough({
         highWaterMark: 128000,
     });
     child.stdout.pipe(pt);
-    return pt;
+    return [pt, () => {
+        stream.unpipe();
+        stream.destroy();
+        cleanup();
+    }];
 }
 
 export {
